Refactor Gene + Artist to use the same Related Artists component
