+++ lib/components/artist/about.js
@@ -7,7 +7,7 @@ import { View, StyleSheet } from 'react-native'
 
 import Biography from './biography'
 import Articles from './articles'
-import RelatedArtists from './related_artists'
+import RelatedArtists from '../related_artists'
 import Separator from '../separator'
 
 class About extends React.Component {
