+++ lib/components/related_artists/index.js
@@ -5,7 +5,7 @@ import Relay from 'react-relay'
 import React from 'react'
 import { StyleSheet, View, Dimensions } from 'react-native'
 
-import SerifText from '../../text/serif'
+import SerifText from '../text/serif'
 import RelatedArtist from './related_artist'
 
 class RelatedArtists extends React.Component {
