+++ lib/components/related_artists/related_artist.js
@@ -5,9 +5,9 @@ import Relay from 'react-relay'
 import React from 'react'
 import { StyleSheet, View, Text, TouchableWithoutFeedback } from 'react-native'
 
-import colors from '../../../../data/colors'
-import ImageView from '../../opaque_image_view'
-import SwitchBoard from '../../../native_modules/switch_board'
+import colors from '../../../data/colors'
+import ImageView from '../opaque_image_view'
+import SwitchBoard from '../../native_modules/switch_board'
 
 class RelatedArtist extends React.Component {
   handleTap() {
